MQTT_GUI_V2/requirements.txt
tk-tools
ttkbootstrap
matplotlib
paho-mqtt